% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_incidence.R
\name{accumulate_incidence}
\alias{accumulate_incidence}
\title{Accumulate incidence from interaction data frame}
\usage{
accumulate_incidence(data, group, incidence = incidence)
}
\arguments{
\item{data}{A \code{data.frame} object with interaction information}

\item{group}{A group column in data \code{data.frame} to accumulate incidence
in an interaction \code{data.frame}}

\item{incidence}{The incidence of interaction in the data \code{data.frame}}
}
\value{
A \code{data.frame} with the summarized incidence from an
interaction \code{data.frame}.
Includes the accumulated incidence for each group provided
}
\description{
Accumulate incidence from interaction data frame
}
\examples{
library(ecointeraction)
accumulate_incidence(mammalvirus, virus, incidence)
}
