% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automodel_replication.R
\name{automodel_replication}
\alias{automodel_replication}
\title{This function returns a list with models given a data frame with}
\usage{
automodel_replication(listData, distance_df)
}
\arguments{
\item{listData}{A list with a sample of labeled species and distance
to predict possible interactions}

\item{distance_df}{A data frame with mean phylogenetic geographical and
environmental distance from one target host species to other species.}
}
\value{
A list with the best fitted model, score of variable importance,
score of ROC-AUC value, score of accuracy, and predicted values for
each possible host.
}
\description{
This function returns a list with models given a data frame with
}
\examples{
if(FALSE){
automodel_replication(listData, birdsdistance_raw)
}
}
