% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{data_example}
\format{
A data frame with 79 rows and 5 variables:
\describe{
  \item{species}{Bird species}
  \item{geo_distance}{Geographical distance}
  \item{env_distance}{Environmental distance}
  \item{phylo_distance}{Phylogenetic distance}
  \item{incidence}{Label if a host has been reported with the interaction}
}
}
\usage{
data_example
}
\description{
A dataset containing the median phylogenetic, geographical and enviromental
distances of birds.
The raw-data has the distances for each pair of birds. This is a sample
ready for modelling. It contains:
Environmental information from WorldClim project
Phylogenetic information from Vertlife project
Geografical information from GBIF project
}
\examples{
data(data_example)
}
\keyword{datasets}
